+++ tests/custom_validators_from_generators_test.go
@@ -1,6 +1,9 @@
 package tests
 
-import "testing"
+import (
+	"context"
+	"testing"
+)
 
 func TestCustomValidatorsFromGenerators(t *testing.T) {
 	scenarios := []scenario{
@@ -24,7 +27,7 @@ func TestCustomValidatorsFromGenerators(t *testing.T) {
 		t.Run(s.name, func(t *testing.T) {
 			defer deferTestCases(t, s.panic, s.expected)
 
-			output := s.validator.Validate(s.in)
+			output := s.validator.Validate(context.TODO(), s.in)
 			check(t, s.expected, output)
 		})
 	}