+++ tests/non_zero_test.go
@@ -1,6 +1,7 @@
 package tests
 
 import (
+	"context"
 	"testing"
 )
 
@@ -75,7 +76,7 @@ func TestNonZero(t *testing.T) {
 		t.Run(s.name, func(t *testing.T) {
 			defer deferTestCases(t, s.panic, s.expected)
 
-			output := s.validator.Validate(s.in)
+			output := s.validator.Validate(context.TODO(), s.in)
 			check(t, s.expected, output)
 		})
 	}