+++ tests/utils.go
@@ -1,6 +1,7 @@
 package tests
 
 import (
+	"context"
 	"testing"
 
 	"github.com/golodash/galidator"
@@ -16,8 +17,8 @@ type scenario struct {
 }
 
 var g = galidator.New().CustomValidators(galidator.Validators{
-	"custom_validator":        func(i interface{}) bool { return false },
-	"second_custom_validator": func(i interface{}) bool { return false },
+	"custom_validator":        func(ctx context.Context, i interface{}) bool { return false },
+	"second_custom_validator": func(ctx context.Context, i interface{}) bool { return false },
 }).CustomMessages(galidator.Messages{"custom_validator": "custom error", "second_custom_validator": "second custom error"})
 
 // Used in test cases to prevent code breaking